+++ packages/aws-amplify-react-native/src/Auth/SignUp.tsx
@@ -32,6 +32,7 @@ import signUpWithUsernameFields, {
 } from './common/default-sign-up-fields';
 import TEST_ID from '../AmplifyTestIDs';
 import { ISignUpField } from '../../types';
+import { setTestId } from '../Utils'
 
 const logger = new Logger('SignUp');
 
@@ -244,7 +245,10 @@ export default class SignUp extends AuthPiece<ISignUpProps, ISignUpState> {
 		this.sortFields();
 		return (
 			<Wrapper>
-				<ScrollView style={theme.sectionScroll}>
+				<ScrollView 
+					style={theme.sectionScroll}
+					keyboardShouldPersistTaps='handled'
+					>
 					<Header theme={theme} testID={TEST_ID.AUTH.SIGN_UP_TEXT}>
 						{I18n.get(this.header)}
 					</Header>
@@ -265,7 +269,7 @@ export default class SignUp extends AuthPiece<ISignUpProps, ISignUpState> {
 									label={I18n.get(field.label)}
 									placeholder={I18n.get(field.placeholder)}
 									required={field.required}
-									testID={field.testID}
+									{...setTestId(field.testID)}
 								/>
 							) : (
 								<PhoneField
@@ -277,7 +281,7 @@ export default class SignUp extends AuthPiece<ISignUpProps, ISignUpState> {
 									keyboardType="phone-pad"
 									required={field.required}
 									defaultDialCode={this.getDefaultDialCode()}
-									testID={field.testID}
+									{...setTestId(field.testID)}
 								/>
 							);
 						})}
@@ -286,7 +290,7 @@ export default class SignUp extends AuthPiece<ISignUpProps, ISignUpState> {
 							theme={theme}
 							onPress={this.signUp}
 							disabled={!this.isValid()}
-							testID={TEST_ID.AUTH.SIGN_UP_BUTTON}
+							{...setTestId(TEST_ID.AUTH.SIGN_UP_BUTTON)}
 						/>
 					</View>
 					<View style={theme.sectionFooter}>